1. Move the unit test about HTTP response of mock APIs to module 'server.application'.
2. Add unit test for function 'create_apis'.
