+++ test/unit_test/server/application.py
@@ -1,15 +1,37 @@
+import json
+import os
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
 from typing import Any, Type
-from unittest.mock import patch
+from unittest.mock import mock_open, patch
 
 import pytest
 from fastapi import FastAPI
 from flask import Flask
 
-from pymock_api.server.application import BaseAppServer, FastAPIServer, FlaskServer
+from pymock_api.exceptions import FileFormatNotSupport
+from pymock_api.server.application import (
+    BaseAppServer,
+    FastAPIServer,
+    FlaskServer,
+    _HTTPResponse,
+)
 
 MockerModule = namedtuple("MockerModule", ["module_path", "return_value"])
+
+_General_String_Value: str = "This is test message for PyTest."
+_Json_File_Name: str = "test.json"
+_Json_File_Content: dict = {"responseCode": "200", "errorMessage": "OK", "content": "This is YouTube home."}
+_Not_Exist_File_Name: str = "not_exist.json"
+_Not_Json_File_Name: str = "test.yaml"
+
+
+class FakeFlask(Flask):
+    pass
+
+
+class FakeFastAPI(FastAPI):
+    pass
 
 
 class AppServerTestSpec(metaclass=ABCMeta):
@@ -23,10 +45,6 @@ class AppServerTestSpec(metaclass=ABCMeta):
     def expected_sut_type(self) -> Any:
         pass
 
-    @abstractmethod
-    def run_target_function(self, sut: BaseAppServer) -> Any:
-        pass
-
     @property
     @abstractmethod
     def mocker(self) -> MockerModule:
@@ -34,19 +52,11 @@ class AppServerTestSpec(metaclass=ABCMeta):
 
     def test_generating_instance_function(self, sut: BaseAppServer):
         with patch(self.mocker.module_path, return_value=self.mocker.return_value) as instantiate_ps:
-            web_app = self.run_target_function(sut)
+            web_app = sut.setup()
             instantiate_ps.assert_called_once()
             assert isinstance(
                 web_app, self.expected_sut_type
             ), f"The web application server it generates should be *{self.expected_sut_type}* type object."
-
-
-class FakeFlask(Flask):
-    pass
-
-
-class FakeFastAPI(FastAPI):
-    pass
 
 
 class TestFlaskServer(AppServerTestSpec):
@@ -62,9 +72,6 @@ class TestFlaskServer(AppServerTestSpec):
     def mocker(self) -> MockerModule:
         return MockerModule(module_path="flask.Flask", return_value=FakeFlask("PyTest-Used"))
 
-    def run_target_function(self, sut: FlaskServer) -> Flask:
-        return sut.setup()
-
 
 class TestFastAPIServer(AppServerTestSpec):
     @pytest.fixture(scope="function")
@@ -79,5 +86,50 @@ class TestFastAPIServer(AppServerTestSpec):
     def mocker(self) -> MockerModule:
         return MockerModule(module_path="fastapi.FastAPI", return_value=FakeFastAPI())
 
-    def run_target_function(self, sut: FastAPIServer) -> Flask:
-        return sut.setup()
+
+class TestInnerHTTPResponse:
+    @pytest.fixture(scope="function")
+    def http_resp(self) -> Type[_HTTPResponse]:
+        return _HTTPResponse
+
+    def test_response_with_string_data(self, http_resp: Type[_HTTPResponse]):
+        resp_data = http_resp.generate(data=_General_String_Value)
+        assert resp_data == _General_String_Value, ""
+
+    def test_response_with_json_format_string_data(self, http_resp: Type[_HTTPResponse]):
+        json_content_str = json.dumps(_Json_File_Content)
+        resp_data = http_resp.generate(data=json_content_str)
+        assert resp_data == json.loads(json_content_str), ""
+
+    def test_response_with_json_file_name(self, http_resp: Type[_HTTPResponse]):
+        with patch.object(os.path, "exists", return_value=True) as os_path_exists:
+            json_content_str = json.dumps(_Json_File_Content)
+            with patch("builtins.open", mock_open(read_data=json_content_str)) as mock_file_stream:
+                # Run target function to test
+                resp_data = http_resp.generate(data=_Json_File_Name)
+
+                # Verify result
+                os_path_exists.assert_called_once_with(_Json_File_Name)
+                mock_file_stream.assert_called_once_with(_Json_File_Name, "r", encoding="utf-8")
+                assert open(_Json_File_Name).read() == json_content_str, ""
+                assert resp_data == _Json_File_Content, ""
+
+    def test_response_with_not_exist_json_file_name(self, http_resp: Type[_HTTPResponse]):
+        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
+            with pytest.raises(FileNotFoundError) as exc_info:
+                # Run target function to test
+                http_resp.generate(data=_Not_Exist_File_Name)
+                # Verify result
+                expected_err_msg = f"The target configuration file {_Not_Exist_File_Name} doesn't exist."
+                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
+                mock_file_stream.assert_not_called()
+
+    def test_response_with_not_json_file_name(self, http_resp: Type[_HTTPResponse]):
+        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
+            with pytest.raises(FileFormatNotSupport) as exc_info:
+                # Run target function to test
+                http_resp.generate(data=_Not_Json_File_Name)
+                # Verify result
+                expected_err_msg = f"It doesn't support reading '{', '.join(http_resp.valid_file_format)}' format file."
+                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
+                mock_file_stream.assert_not_called()
