+++ test/unit_test/server/mock.py
@@ -1,20 +1,23 @@
-import json
-import os
-from typing import Callable, Type
-from unittest.mock import Mock, mock_open, patch
+from typing import Any, Callable
+from unittest.mock import Mock, patch
 
 import pytest
 
-from pymock_api.exceptions import FileFormatNotSupport
 from pymock_api.model.api_config import APIConfig, MockAPIs
-from pymock_api.server.application import FlaskServer
-from pymock_api.server.mock import MockHTTPServer, _HTTPResponse
+from pymock_api.server.application import BaseAppServer, FlaskServer
+from pymock_api.server.mock import MockHTTPServer
 
-_General_String_Value: str = "This is test message for PyTest."
-_Json_File_Name: str = "test.json"
-_Json_File_Content: dict = {"responseCode": "200", "errorMessage": "OK", "content": "This is YouTube home."}
-_Not_Exist_File_Name: str = "not_exist.json"
-_Not_Json_File_Name: str = "test.yaml"
+from ..._values import _Test_Config
+
+mock_api_config = Mock(APIConfig(name=Mock(), description=Mock(), apis=Mock(MockAPIs(base=Mock(), apis=Mock()))))
+
+
+class FakeWebServer(BaseAppServer):
+    def setup(self) -> Any:
+        pass
+
+    def create_api(self, mocked_apis: MockAPIs) -> None:
+        pass
 
 
 class TestMockHTTPServer:
@@ -72,6 +75,15 @@ class TestMockHTTPServer:
             instantiate_callback=_instantiate, assert_config_path="api.yaml", auto_setup=True
         )
 
+    @patch("pymock_api.server.mock.load_config", return_value=mock_api_config)
+    @patch.object(FakeWebServer, "create_api")
+    def test_create_apis(self, mock_create_apis: Mock, mock_load_config: Mock):
+        apis = Mock(MockAPIs(base=Mock(), apis=Mock()))
+        mock_server = MockHTTPServer(config_path=_Test_Config, app_server=FakeWebServer(), auto_setup=False)
+        mock_server.create_apis(apis)
+        mock_create_apis.assert_called_once_with(apis)
+        mock_load_config.assert_called_once_with(config_path=_Test_Config)
+
     @staticmethod
     def _template_test(
         instantiate_callback: Callable,
@@ -118,51 +130,3 @@ class TestMockHTTPServer:
                         mock_app_server=mock_flask_server_obj,
                         instantiate_flask_app_server=True,
                     )
-
-
-class TestInnerHTTPResponse:
-    @pytest.fixture(scope="function")
-    def http_resp(self) -> Type[_HTTPResponse]:
-        return _HTTPResponse
-
-    def test_response_with_string_data(self, http_resp: Type[_HTTPResponse]):
-        resp_data = http_resp.generate(data=_General_String_Value)
-        assert resp_data == _General_String_Value, ""
-
-    def test_response_with_json_format_string_data(self, http_resp: Type[_HTTPResponse]):
-        json_content_str = json.dumps(_Json_File_Content)
-        resp_data = http_resp.generate(data=json_content_str)
-        assert resp_data == json.loads(json_content_str), ""
-
-    def test_response_with_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch.object(os.path, "exists", return_value=True) as os_path_exists:
-            json_content_str = json.dumps(_Json_File_Content)
-            with patch("builtins.open", mock_open(read_data=json_content_str)) as mock_file_stream:
-                # Run target function to test
-                resp_data = http_resp.generate(data=_Json_File_Name)
-
-                # Verify result
-                os_path_exists.assert_called_once_with(_Json_File_Name)
-                mock_file_stream.assert_called_once_with(_Json_File_Name, "r", encoding="utf-8")
-                assert open(_Json_File_Name).read() == json_content_str, ""
-                assert resp_data == _Json_File_Content, ""
-
-    def test_response_with_not_exist_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
-            with pytest.raises(FileNotFoundError) as exc_info:
-                # Run target function to test
-                http_resp.generate(data=_Not_Exist_File_Name)
-                # Verify result
-                expected_err_msg = f"The target configuration file {_Not_Exist_File_Name} doesn't exist."
-                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
-                mock_file_stream.assert_not_called()
-
-    def test_response_with_not_json_file_name(self, http_resp: Type[_HTTPResponse]):
-        with patch("builtins.open", mock_open(read_data=None)) as mock_file_stream:
-            with pytest.raises(FileFormatNotSupport) as exc_info:
-                # Run target function to test
-                http_resp.generate(data=_Not_Json_File_Name)
-                # Verify result
-                expected_err_msg = f"It doesn't support reading '{', '.join(http_resp.valid_file_format)}' format file."
-                assert str(exc_info) == expected_err_msg, f"The error message should be same as '{expected_err_msg}'."
-                mock_file_stream.assert_not_called()
